reading the price

+++ BikeShop1/src/Model/BikeModel.java
@@ -49,7 +49,7 @@ public class BikeModel {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
